requirements.txt
Django>=3.0.4
djangorestframework>=3.11.0
coverage>=5.0.4